pi_dma/pi_util: Fix up/clarify bus/phys addresses
On the BCM2835, the DMA can use the L2 cache, on BCM2836/2837 it cannot.
So, fix board detection to make sure the correct memory attributes are
used (CPU mapping should use L2 on BCM2835), and rename
phys_virt_to_phys to phys_virt_to_bus to reflect the fact that it's
actually returning bus addresses (which might use the L2 or might not).

Similarly, clarify naming on other "physical" addresses.