pi_backend: Plumb through the fence wait parameters
So that they aren't hard-coded magic in the backend code, pass in the
timeout and the sleep time from the top level.

timeout_millis should be set to something large to detect underruns.

sleep_millis should be set as large as possible (25-50% of chunk time)
to not keep the CPU too busy polling.